% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhcf.pq-package.R
\docType{package}
\name{bhcf.pq-package}
\alias{bhcf.pq}
\alias{bhcf.pq-package}
\title{bhcf.pq: FFIEC BHCF Processing Tools}
\description{
Tools for discovering, downloading, parsing, and processing FFIEC BHCF bulk
data, including conversion to Parquet format and integration with DuckDB.
}
\author{
\strong{Maintainer}: Ian Gow \email{iandgow@gmail.com}

}
\keyword{internal}
