% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhcf_manifest.R
\name{bhcf_check_pq_keys}
\alias{bhcf_check_pq_keys}
\title{Check primary-key and non-NULL constraints in BHCF Parquet files}
\usage{
bhcf_check_pq_keys(
  conn,
  cols,
  pq_file = NULL,
  data_dir = NULL,
  schema = "bhcf"
)
}
\arguments{
\item{conn}{A valid DuckDB connection.}

\item{cols}{Character vector of columns that should be non-NULL and unique
jointly.}

\item{pq_file}{Optional Parquet filename/glob passed to [bhcf_scan_pqs()].}

\item{data_dir}{Optional parent directory for Parquet output.}

\item{schema}{Schema subdirectory. Default is `"bhcf"`.}
}
\value{
A tibble with key-check results.
}
\description{
Check primary-key and non-NULL constraints in BHCF Parquet files
}
